Split constants
