+++ core/src/com/po/fuck/model/Coins.java
@@ -3,8 +3,8 @@ package com.po.fuck.model;
 import com.badlogic.gdx.math.Vector2;
 import com.po.fuck.view.Drawable;
 
-import static com.po.fuck.model.Constants.COINS_COUNTER_POSITION;
-import static com.po.fuck.model.Constants.COIN_SPRITE_POSITION;
+import static com.po.fuck.view.Constants.COINS_COUNTER_POSITION;
+import static com.po.fuck.view.Constants.COIN_SPRITE_POSITION;
 import static com.po.fuck.model.Constants.GUI_LAYER;
 
 public class Coins implements Drawable {
