+++ core/src/com/po/fuck/view/FollowingDrawer.java
@@ -1,7 +1,7 @@
 package com.po.fuck.view;
 
-import static com.po.fuck.model.Constants.GAME_HEIGHT;
-import static com.po.fuck.model.Constants.GAME_WIDTH;
+import static com.po.fuck.view.Constants.GAME_HEIGHT;
+import static com.po.fuck.view.Constants.GAME_WIDTH;
 
 import com.badlogic.gdx.graphics.g2d.Sprite;
 import com.badlogic.gdx.graphics.g2d.SpriteBatch;
