--- core/src/com/po/fuck/view/ViewConstants.java
@@ -1,8 +1,0 @@
-package com.po.fuck.view;
-
-public class ViewConstants {
-    /* health bar */
-    public static final float HEALTHBAR_OFFSET = 5;
-    public static final int HEALTHBAR_WIDTH = 100;
-    public static final int HEALTHBAR_HEIGHT = 10;
-}
